move bar transform functions from bar/plot to bar/helpers
